#21: Added slog instead of log | moved pkg/util/cmds to /util/cmds
